refactor(js-toolkit-core)!: make room for liferay-cli project descriptor
