+++ projects/js-toolkit/packages/js-toolkit-core/src/project/bundler3/Project.ts
@@ -11,10 +11,10 @@ import readJsonSync from 'read-json-sync';
 import resolveModule from 'resolve';
 import webpack from 'webpack';
 
-import FilePath from '../file/FilePath';
-import {info, print, warn} from '../format';
-import {splitModuleName} from '../node/modules';
-import PkgJson from '../schema/PkgJson';
+import FilePath from '../../file/FilePath';
+import {info, print, warn} from '../../format';
+import {splitModuleName} from '../../node/modules';
+import PkgJson from '../../schema/PkgJson';
 import Adapt from './Adapt';
 import Jar from './Jar';
 import Localization from './Localization';
