+++ projects/js-toolkit/packages/js-toolkit-core/src/project/bundler3/VersionInfo.ts
