+++ projects/js-toolkit/packages/js-toolkit-core/src/project/bundler3/__tests__/__fixtures__/default-features/liferay-npm-bundler.config.js
