+++ projects/js-toolkit/packages/js-toolkit-core/src/project/bundler3/__tests__/__fixtures__/project/bool-create-jar/liferay-npm-bundler.config.js
