+++ projects/js-toolkit/packages/js-toolkit-core/src/project/bundler3/__tests__/__fixtures__/project/create-jar-off-with-portlet/liferay-npm-bundler.config.js
