+++ projects/js-toolkit/packages/js-toolkit-core/src/project/bundler3/getFeaturesFilePath.ts
@@ -6,7 +6,7 @@
 import prop from 'dot-prop';
 import fs from 'fs';
 
-import FilePath from '../file/FilePath';
+import FilePath from '../../file/FilePath';
 import Project from './Project';
 
 /**
